+++ Sources/ChrisDataStructures/Lists/Nodes/DoublyLinkedNode.swift
