+++ Sources/ChrisDataStructures/Lists/Nodes/ThreeDirectionNode.swift
